Develop (#1)
* optimization : change SYNC_MASTER flushSlave/flushDisk to pipeline to reduce produce latency and improve throughput

* WIP: compatible with JDK1.6

* compatible with JDK1.6 in remoting module

* fix checkStyle

* compatible with dledger

* compatible with dledger

* 1. eliminate redundant CountDownLaunch in GroupCommitRequest
2. fix properties when store message
3. add IT for transaction
4. correct resonse code when slave down

* Fix asynchronous send retry

* Improve exception handle of transaction check service (#1783)

* [ISSUE #1787] Fix the issue that queryCq command of mqadmin returned incorrect data (#1793)

* fix(tools):fix the issue that queryCq command of mqadmin returned incorrect data

* fix(mqadmin):set status code in advance in queryConsumeQueue method

* chore(pom):update fastjson version (#1795)

* fix(pull_schedual) add the namespace wrapper in callback (#1804)

* add delete IndexFile

* chore(version):modify current version to prepare release 4.7.0

* chore(all):update the year in notice file and correct header in NamespaceUtilTest

* fix(test):fix the issue that MessageClientIDSetter#getIPFromID return wrong pid when pid greater than short type max value

* [maven-release-plugin] prepare release rocketmq-all-4.7.0

* [maven-release-plugin] prepare for next development iteration

* [ISSUE #1857] Refactor the duplicated code

close #1857

* fix transaction message doc

* feat(broker) prevent blocked by group transfer service

* fix #1883
fix wipeWritePerm doc desc

* [ISSUE #1741] Rename takeMessags to takeMessages

Closes #1741

* [ISSUE #1505]Fix the ListSplitter of batch sample in docs

* [ISSUE #1574] change GET_ROUTEINTO_BY_TOPIC to GET_ROUTEINFO_BY_TOPIC

[ISSUE #1574] change GET_ROUTEINTO_BY_TOPIC to GET_ROUTEINFO_BY_TOPIC

* [ISSUE#728]check ChannelHandlerContext whether is null

[ISSUE#728]check ChannelHandlerContext whether is null

* Correct the Java doc for MessageListenerOrderly.java (#1897)

* Update the javadoc for MessageListenerOrderly

* [ISSUE #1859] add codecov report for unit tests.

[ISSUE #1859] add codecov report for tests.

* [ISSUE #1866] Fix typo in document

* [ISSUE #1895] Expose the flush disk timeout error

* [ISSUE #1901] Fix bug in sending request/reply message using broker 4.7.0

fix(broker):fix bug sending request message using broker 4.7.0 (#1902)

* [ISSUE #1463] Skip encode the null value property as string "null" (#1899)

* Update MessageDecoder.java
* add test case for MessageDecoder

* [ISSUE #1912]Polish the committed offset logic for the lite pull consumer.

[ISSUE #1912]Polish the committed offset logic for the lite pull consumer.

* fixed typo in RequestResponseFuture.
rename RequestResponseFuture.setSendReqeustOk to
RequestResponseFuture.setSendRequestOk.

* fix #1957
DefaultMQProducer didn't shutdown completely

* remove duplicate shutdown producer

* [ISSUE #1964]fix Doc error: Connection related command instructions

Doc error: Connection related command instructions

* [ISSUE #1906] The BooleanConstantExpression might lead to class loading deadlock

[ISSUE #1906] The BooleanConstantExpression might lead to class loading deadlock

* add request-reply doc

* modify request-reply doc

* modify request-reply doc

* modify request-reply doc

* modify request-reply doc

* fix Broker2Client logs

* modify request-reply doc

* Fastjson has a serious security problem in 1.2.62,which will cause RCE

Fastjson has a serious security problem in 1.2.62,which will cause RCE

Fix https://github.com/apache/rocketmq/issues/2007

* fix issue #705 (#706)

+++ common/src/test/java/org/apache/rocketmq/common/message/MessageClientIDSetterTest.java
@@ -40,7 +40,7 @@ public class MessageClientIDSetterTest {
         int pid = UtilAll.getPid();
 
         String uniqID = MessageClientIDSetter.createUniqID();
-        short pidFromID = MessageClientIDSetter.getPidFromID(uniqID);
+        int pidFromID = MessageClientIDSetter.getPidFromID(uniqID);
 
         assertThat(pid).isEqualTo(pidFromID);
     }
